// src/types/types-1.36.3.ts
export interface VaultExplorerPluginSettings_1_36_3 {
	properties: {
		favorite: string;
		url: string;
		imageUrl: string;
		createdDate: string;
		modifiedDate: string;
		custom1: string;
		custom2: string;
		custom3: string;
	};
	filters: {
		search: TSearchFilter;
		favorites: TFavoritesFilter;
		sort: TSortFilter;
		timestamp: TTimestampFilter;
		custom: TCustomFilter;
	};
	views: {
		dashboard: TDashboardView;
		grid: TGridView;
		list: TListView;
		table: TTableView;
		feed: TFeedView;
		recommended: TRecommendedView;
		related: TRelatedView;
	};
	titleWrapping: WordBreak;
	enableClockUpdates: boolean;
	enableFileIcons: boolean;
	currentView: TExplorerView | null;
	enableScrollButtons: boolean;
	pageSize: number;
	filterGroupsWidth: string;
	shouldWrapFilterGroups: boolean;
	viewOrder: TExplorerView[];
	configDir: string;
	pluginVersion: string | null;
	logLevel: string;
}

type FileInteractionStyle = "title" | "content";

interface BaseView {
	isEnabled: boolean;
}

interface TTableView extends BaseView {}

interface TListView extends BaseView {
	showTags: boolean;
}

interface TGridView extends BaseView {
	coverImageSource: CoverImageSource;
	loadSocialMediaImage: boolean;
}

type CoverImageSource = "frontmatter-only" | "frontmatter-and-body" | "off";

interface TDashboardView extends BaseView {}

type CollapseStyle = "no-new-lines" | "no-extra-new-lines";

interface TFeedView extends BaseView {
	collapseStyle: CollapseStyle;
	removeH1: boolean;
	lineClampSmall: number;
	lineClampMedium: number;
	lineClampLarge: number;
}

interface TRecommendedView extends BaseView {}

interface TRelatedView extends BaseView {}

interface BaseFilter {
	isEnabled: boolean;
}

interface TSearchFilter extends BaseFilter {
	value: string;
}

interface TFavoritesFilter extends BaseFilter {
	value: boolean;
}

interface TSortFilter extends BaseFilter {
	value: SortFilterOption;
}

interface TTimestampFilter extends BaseFilter {
	value: TimestampFilterOption;
}

interface TCustomFilter extends BaseFilter {
	selectedGroupId: string;
	groups: TFilterGroup[];
}

type TimestampFilterOption =
	| "created-today"
	| "modified-today"
	| "created-this-week"
	| "modified-this-week"
	| "created-2-weeks"
	| "modified-2-weeks"
	| "all";

type SortFilterOption =
	| "file-name-asc"
	| "file-name-desc"
	| "modified-asc"
	| "modified-desc"
	| "random";

type WordBreak = "normal" | "break-word";

enum TExplorerView {
	DASHBOARD = "dashboard",
	GRID = "grid",
	LIST = "list",
	FEED = "feed",
	TABLE = "table",
	RECOMMENDED = "recommended",
	RELATED = "related",
}
// ...
interface TFilterGroup {
	id: string;
	name: string;
	rules: TFilterRule[];
	isEnabled: boolean;
	isSticky: boolean;
}